+++ src/components/pokemon-list/__test__/pokemon-list.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen } from '@testing-library/react';
 import { MemoryRouter } from 'react-router';
 import PokemonList from '../pokemon-list';
-import type { Pokemon } from '../../types/pokemon';
+import type { Pokemon } from '../../shared/types/pokemon';
 
 describe('PokemonList', () => {
   const pokemons: Pokemon[] = [
