+++ src/components/pokemon-list/components/pokemon-details-card/__test__/pokemon-details-card.test.tsx
@@ -1,9 +1,9 @@
 import { render, screen, waitFor, fireEvent } from '@testing-library/react';
 import { MemoryRouter, Route, Routes } from 'react-router';
-import PokemonDetailsCard from '../pokemon-details-card';
-import { getPokemonDetails } from '../../service/pokemon-service';
+import { PokemonDetailsCard } from '../pokemon-details-card';
+import { getPokemonDetails } from '../../../../../api/pokemon-api/pokemon-service';
 
-import type { PokemonDetails } from '../../types/pokemon-details';
+import type { PokemonDetails } from '../types/pokemon-details';
 import { vi } from 'vitest';
 import type { Mock } from 'vitest';
