+++ src/components/pokemon-list/components/pokemon-details-card/types/pokemon-details.ts
