+++ src/pages/search-page/__tests__/search-page.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen, fireEvent, waitFor } from '@testing-library/react';
 import { vi } from 'vitest';
-import SearchPage from '../search-page';
+import { SearchPage } from '../search-page';
 import { MemoryRouter } from 'react-router';
 
 type Pokemon = {
