+++ src/router/router.ts
@@ -1,8 +1,8 @@
 import { createBrowserRouter } from 'react-router';
-import SearchPage from '../components/search-page';
-import NotFoundPage from '../components/not-found-page/not-found-page';
-import PokemonDetailsCard from '../components/pokemon-list/pokemon-details-card';
-import AboutPage from '../components/about-page/about';
+import { SearchPage } from '../pages/search-page/search-page';
+import { NotFoundPage } from '../pages/not-found-page/not-found-page';
+import { PokemonDetailsCard } from '../components/pokemon-list/components/pokemon-details-card/pokemon-details-card';
+import { AboutPage } from '../pages/about-page/about';
 import { ROUTES_PATH } from './routes-path';
 
 export const router = createBrowserRouter([
