+++ src/types/string-nullable.ts
@@ -1,0 +1,1 @@
+export type StringNullable = string | null;
